refactored day2 to use intcomputer

+++ src/main/java/be/selske/aoc2019/days/Day5.java
@@ -26,7 +26,7 @@ public class Day5 extends AocDay {
         return runProgram(5, parseInput(input));
     }
 
-    private static String runProgram(int inputValue, long[] inputs) {
+    private static String runProgram(long inputValue, long[] inputs) {
         IntComputer intComputer = new IntComputer(Arrays.copyOf(inputs, inputs.length));
 
         intComputer.run(inputValue);
